+++ packages/fiori-app-sub-generator/test/int/fiori-freestyle/expected-output/simple_no_datasource/webapp/test/unit/AllTests.js
@@ -1,0 +1,5 @@
+sap.ui.define([
+	"sap/com/simplenodatasource/test/unit/controller/View1.controller"
+], function () {
+	"use strict";
+});