+++ packages/deploy-config-generator-shared/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/deploy-config-generator-shared
 
+## 0.0.40
+
+### Patch Changes
+
+-   Updated dependencies [23e055a]
+    -   @sap-ux/fiori-generator-shared@0.10.0
+
 ## 0.0.39
 
 ### Patch Changes